adding correct update_for_pi. using float64 for higher precision

+++ OpenCL/sampler.cl
@@ -37,14 +37,14 @@ int sample_z_ab_from_edge(
 
 // merged version
 int sample_z_ab_from_edge(
-		global float* pi_a,
-		global float *pi_b,
-		global float *beta,
-		float epsilon, float y,
-		global float *p,
-		global float *bounds) {
-	float location = 0;
-	float bound = pow(beta[0], y) * pow(1-beta[0], 1-y) * pi_a[0] * pi_b[0]
+		global double* pi_a,
+		global double *pi_b,
+		global double *beta,
+		double epsilon, double y,
+		global double *p,
+		global double *bounds) {
+	double location = 0;
+	double bound = pow(beta[0], y) * pow(1-beta[0], 1-y) * pi_a[0] * pi_b[0]
 		+ pow(epsilon, y) * pow(1-epsilon, 1-y) * pi_a[0] * (1-pi_b[0]);
 	if (location <= bound) return 0;
 	for (int i = 1; i < K; ++i) {
@@ -61,12 +61,12 @@ void sample_latent_vars_of(
 		int node,
 		global Graph *g,
 		global int* neighbor_nodes,
-		global float *pi,
-		global float *beta,
-		float epsilon,
-		global float *z /* K elements */,
-		global float *p,
-		global float *bounds) {
+		global double *pi,
+		global double *beta,
+		double epsilon,
+		global double *z /* K elements */,
+		global double *p,
+		global double *bounds) {
 	for (int i = 0; i < K; ++i) z[i] = 0;
 	for (int i = 0; i < NEIGHBOR_SAMPLE_SIZE; ++i) {
 		int neighbor = neighbor_nodes[i];
@@ -83,17 +83,17 @@ kernel void sample_latent_vars(
 		global int *nodes,
 		int N, // #nodes
 		global int *neighbor_nodes,// (#total_nodes, NEIGHBOR_SAMPLE_SIZE)
-		global float *pi,// (#total_nodes, K)
-		global float *beta,// (#K)
-		float epsilon,
-		global float *Z,// (#total_nodes, K)
-		global float *p,// (#nodes, K)
-		global float *bounds// (#nodes, K)
+		global double *pi,// (#total_nodes, K)
+		global double *beta,// (#K)
+		double epsilon,
+		global double *Z,// (#total_nodes, K)
+		global double *p,// (#nodes, K)
+		global double *bounds// (#nodes, K)
 ) {
 	size_t gid = get_global_id(0);
 	size_t gsize = get_global_size(0);
-	global float *_p = p + gid * K;
-	global float *_b = bounds + gid * K;
+	global double *_p = p + gid * K;
+	global double *_b = bounds + gid * K;
 
 	for (int i = gid; i < N; i += gsize) {
 		int node = nodes[i];
@@ -111,29 +111,30 @@ kernel void sample_latent_vars(
 
 void update_pi_for_node_(
 		int node,
-		global float *pi,// #K
-		global float *phi,// #K
-		global float *z, // #K
-		global float *noise, // #K
-		global float *grad, // #K
-		float alpha,
-		float a, float b, float c,
-		float step_count, int total_node_count
+		global double *pi,// #K
+		global double *phi,// #K
+		global double *z, // #K
+		global double *noise, // #K
+		global double *grad, // #K
+		double alpha,
+		double a, double b, double c,
+		int step_count, int total_node_count
 		) {
-	float eps_t = a * pow((1 + step_count/b), -c);
-	float phi_i_sum = 0;
+	double eps_t = a * pow((1 + step_count/b), -c);
+	double phi_i_sum = 0;
 	for (int i = 0; i < K; ++i) phi_i_sum += phi[i];
 	for (int k = 0; k < K; ++k) {
 		grad[k] = -NEIGHBOR_SAMPLE_SIZE * 1/phi_i_sum;
 		grad[k] += 1/phi[k] * z[k];
 	}
 	for (int k = 0; k < K; ++k) {
-		float phi_star_k = fabs(phi[k] + eps_t/2
+		double phi_star_k = fabs(phi[k] + eps_t/2
 				* (alpha - phi[k] + total_node_count/NEIGHBOR_SAMPLE_SIZE * grad[k])
-				+ pow(eps_t, 0.5f) * pow(phi[k], 0.5f) * noise[k]);
-		phi[k] = phi_star_k * (1.0f/step_count) + (1-1.0f/step_count) * phi[k];
+				+ pow(eps_t, 0.5) * pow(phi[k], 0.5) * noise[k]);
+		phi[k] = phi_star_k * (1.0/step_count)
+				+ (1-1.0/step_count) * phi[k];
 	}
-	float phi_sum = 0;
+	double phi_sum = 0;
 	for (int i = 0; i < K; ++i) phi_sum += phi[i];
 	for (int i = 0; i < K; ++i) {
 		pi[i] = phi[i]/phi_sum;
@@ -143,14 +144,14 @@ void update_pi_for_node_(
 kernel void update_pi_for_node(
 		global int *nodes,
 		int N, // #nodes
-		global float *pi,// (#total_nodes, K)
-		global float *phi,// (#total_nodes, K)
-		global float *Z, // (#total_nodes, K)
-		global float *noise, // (#nodes, K)
-		global float *grad, // (#nodes, K)
-		float alpha,
-		float a, float b, float c,
-		float step_count, int total_node_count
+		global double *pi,// (#total_nodes, K)
+		global double *phi,// (#total_nodes, K)
+		global double *Z, // (#total_nodes, K)
+		global double *noise, // (#nodes, K)
+		global double *grad, // (#nodes, K)
+		double alpha,
+		double a, double b, double c,
+		int step_count, int total_node_count
 		) {
 	size_t gid = get_global_id(0);
 	size_t gsize = get_global_size(0);
